visual: extract HUD mode from vmenus.c to hud.c
