+++ librz/core/tui/vmenus.c
@@ -491,120 +491,6 @@ RZ_IPI bool rz_core_visual_bit_editor(RzCore *core) {
 	return true;
 }
 
-RZ_IPI bool rz_core_visual_hudclasses(RzCore *core) {
-	RzListIter *iter, *iter2;
-	RzBinClass *c;
-	RzBinField *f;
-	RzBinSymbol *m;
-	ut64 addr;
-	char *res;
-	RzList *list = rz_list_new();
-	if (!list) {
-		return false;
-	}
-	list->free = free;
-	RzList *classes = rz_bin_get_classes(core->bin);
-	rz_list_foreach (classes, iter, c) {
-		rz_list_foreach (c->fields, iter2, f) {
-			rz_list_append(list, rz_str_newf("0x%08" PFMT64x "  %s %s", f->vaddr, c->name, f->name));
-		}
-		rz_list_foreach (c->methods, iter2, m) {
-			const char *name = m->dname ? m->dname : m->name;
-			rz_list_append(list, rz_str_newf("0x%08" PFMT64x "  %s %s", m->vaddr, c->name, name));
-		}
-	}
-	res = rz_cons_hud(list, NULL);
-	if (res) {
-		char *p = strchr(res, ' ');
-		if (p) {
-			*p = 0;
-		}
-		addr = rz_num_get(NULL, res);
-		rz_core_seek(core, addr, true);
-		free(res);
-	}
-	rz_list_free(list);
-	return res != NULL;
-}
-
-static bool hudstuff_append(RzFlagItem *fi, void *user) {
-	RzList *list = (RzList *)user;
-	char *s = rz_str_newf("0x%08" PFMT64x "  %s", fi->offset, fi->name);
-	if (s) {
-		rz_list_append(list, s);
-	}
-	return true;
-}
-
-RZ_IPI bool rz_core_visual_hudstuff(RzCore *core) {
-	ut64 addr;
-	char *res;
-	RzList *list = rz_list_new();
-	if (!list) {
-		return false;
-	}
-	list->free = free;
-	rz_flag_foreach(core->flags, hudstuff_append, list);
-	RzIntervalTreeIter it;
-	RzAnalysisMetaItem *mi;
-	rz_interval_tree_foreach (&core->analysis->meta, it, mi) {
-		if (mi->type == RZ_META_TYPE_COMMENT) {
-			char *s = rz_str_newf("0x%08" PFMT64x " %s", rz_interval_tree_iter_get(&it)->start, mi->str);
-			if (s) {
-				rz_list_push(list, s);
-			}
-		}
-	}
-	res = rz_cons_hud(list, NULL);
-	if (res) {
-		char *p = strchr(res, ' ');
-		if (p) {
-			*p = 0;
-		}
-		addr = rz_num_get(NULL, res);
-		rz_core_seek(core, addr, true);
-		free(res);
-	}
-	rz_list_free(list);
-	return res != NULL;
-}
-
-static bool rz_core_visual_config_hud(RzCore *core) {
-	RzListIter *iter;
-	RzConfigNode *bt;
-	RzList *list = rz_list_new();
-	if (!list) {
-		return false;
-	}
-	list->free = free;
-	rz_list_foreach (core->config->nodes, iter, bt) {
-		rz_list_append(list, rz_str_newf("%s %s", bt->name, bt->value));
-	}
-	char *res = rz_cons_hud(list, NULL);
-	if (res) {
-		const char *oldvalue = NULL;
-		char cmd[512];
-		char *p = strchr(res, ' ');
-		if (p) {
-			*p = 0;
-		}
-		oldvalue = rz_config_get(core->config, res);
-		rz_cons_show_cursor(true);
-		rz_cons_set_raw(false);
-		cmd[0] = '\0';
-		eprintf("Set new value for %s (old=%s)\n", res, oldvalue);
-		rz_line_set_prompt(":> ");
-		if (rz_cons_fgets(cmd, sizeof(cmd), 0, NULL) < 0) {
-			cmd[0] = '\0';
-		}
-		rz_config_set(core->config, res, cmd);
-		rz_cons_set_raw(true);
-		rz_cons_show_cursor(false);
-	}
-	rz_list_free(list);
-	return true;
-}
-
 // TODO: skip N first elements
 // TODO: show only N elements of the list
 // TODO: wrap index when out of boundaries
